written delete,post,update rooms methods

+++ src/main/java/com/hotel/management/model/Rooms.java
@@ -1,15 +1,21 @@
 package com.hotel.management.model;
 
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+
+
 public class Rooms {
-    private long id;
+
+    private int id;
     private String date;
     private int roomNo;
     private String availability;
 
     public Rooms() {
     }
-
-    public Rooms(long id, String date, int roomNo, String availability) {
+    public Rooms(int id, String date, int roomNo, String availability) {
         this.id = id;
         this.date = date;
         this.roomNo = roomNo;
@@ -20,7 +26,7 @@ public class Rooms {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
